requirements.txt
pandas==2.2.3
google-cloud-aiplatform==1.87.0
kfp==2.0.0
notebook==7.3.3